downgrade to 1.3.0, as 1.4.0 is not released yet
Since I haven't released 1.4.0 yet (always grab snapblocks from the latest release, don't grab it from the dev site), things can break. Yes, there are some bugs in 1.3.0 that are fixed in 1.4.0, but it's best to stay on the latest stable release (especially when snapblocks is in active development).

I also added the showSpaces option, and the snapblocks version is logged to the console.